refactor(apps/pool): v3 pool page
