+++ apps/evm/app/analytics/components/TableSection/FuroTokens/FuroTokenTable.tsx
@@ -72,7 +72,6 @@ export const FuroTokenTable: FC = () => {
           onPrev={table.previousPage}
           onNext={table.nextPage}
           page={pagination.pageIndex}
-          onPage={table.setPageIndex}
           pages={table.getPageCount()}
           pageSize={PAGE_SIZE}
         />
