+++ apps/evm/app/analytics/components/TableSection/Tokens/TokenTable.tsx
@@ -129,7 +129,6 @@ export const TokenTable: FC = () => {
         onPrev={table.previousPage}
         onNext={table.nextPage}
         page={pagination.pageIndex}
-        onPage={table.setPageIndex}
         pages={table.getPageCount()}
         pageSize={PAGE_SIZE}
       />
