--- apps/evm/ui/pool/PoolSection/V3/PoolDepthWidget/index.ts
@@ -1,1 +1,0 @@
-export * from './PoolDepthWidget'
