--- apps/evm/ui/pool/PoolSection/V3/PoolTransactionsV3/index.ts
@@ -1,1 +1,0 @@
-export * from './PoolTransactionsV3'
