+++ apps/evm/ui/pool/PoolSection/V3/Tabs/Statistics/Charts/Charts.tsx
@@ -1,27 +1,17 @@
 import { SushiSwapV3ChainId } from '@sushiswap/v3-sdk'
 import React, { FC, useMemo, useState } from 'react'
 
-import { PoolDepthWidget } from '../V3/PoolDepthWidget'
-import { PoolChartGraph } from './PoolChartGraph'
-import { PoolChartPeriods } from './PoolChartPeriods'
-import { PoolChartTypes } from './PoolChartTypes'
-import { PoolChartPeriod, PoolChartType } from './types'
+import { PoolChartGraph, PoolChartPeriod, PoolChartPeriods, PoolChartType, PoolChartTypes } from '../../../../PoolChart'
+import { LiquidityDepthWidget } from './LiquidityDepthWidget'
 
 const charts = [PoolChartType.Volume, PoolChartType.TVL, PoolChartType.Fees, PoolChartType.Depth]
-const periods = [
-  PoolChartPeriod.Day,
-  PoolChartPeriod.Week,
-  PoolChartPeriod.Month,
-  PoolChartPeriod.Year,
-  PoolChartPeriod.All,
-]
 
-interface PoolChartV3Props {
+interface Charts {
   address: string
   chainId: SushiSwapV3ChainId
 }
 
-const PoolChartV3: FC<PoolChartV3Props> = ({ address, chainId }) => {
+export const Charts: FC<Charts> = ({ address, chainId }) => {
   const [chart, setChart] = useState<PoolChartType>(charts[0])
   const [period, setPeriod] = useState<PoolChartPeriod>(PoolChartPeriod.Month)
 
@@ -38,12 +28,10 @@ const PoolChartV3: FC<PoolChartV3Props> = ({ address, chainId }) => {
         <PoolChartPeriods periods={periods} selectedPeriod={period} setPeriod={setPeriod} />
       </div>
       {chart === PoolChartType.Depth ? (
-        <PoolDepthWidget chainId={chainId} address={address} />
+        <LiquidityDepthWidget chainId={chainId} address={address} />
       ) : (
         <PoolChartGraph chart={chart} period={period} address={address} chainId={chainId} />
       )}
     </div>
   )
 }
-
-export { PoolChartV3 }
