+++ apps/evm/ui/pool/PoolSection/V3/Tabs/Statistics/Charts/PoolDepthChart.tsx
@@ -4,7 +4,7 @@ import ReactECharts, { EChartsOption } from 'echarts-for-react'
 import React, { FC, useCallback, useMemo } from 'react'
 import colors from 'tailwindcss/colors'
 
-import { ChartEntry } from '../../../LiquidityChartRangeInput/types'
+import { ChartEntry } from '../../../../../LiquidityChartRangeInput/types'
 
 interface PoolDepthChartProps {
   poolStats: NonNullable<ReturnType<typeof useConcentratedLiquidityPoolStats>['data']>
