+++ apps/evm/ui/pool/PoolSection/V3/Tabs/Statistics/Charts/index.ts
@@ -1,0 +1,1 @@
+export { Charts } from './Charts'
