+++ apps/evm/ui/pool/PoolSection/V3/Tabs/Statistics/Transactions/Transactions.tsx
@@ -8,9 +8,9 @@ import { getCoreRowModel, getSortedRowModel, useReactTable } from '@tanstack/rea
 import React, { FC, useMemo, useState } from 'react'
 
 import { AMOUNT_IN_COLUMN, AMOUNT_OUT_COLUMN, AMOUNT_USD_COLUMN, SENDER_COLUMN, TIME_COLUMN } from './columns'
-import { Transaction, TransactionType, useTransactionsV3 } from './useTransactionsV3'
+import { Transaction, TransactionType, useTransactions } from './useTransactions'
 
-interface PoolTransactionsV3Props {
+interface TransactionsProps {
   pool: Pool | undefined | null
   poolId: string
 }
@@ -19,8 +19,8 @@ const TOTAL_ROWS = 200
 const PAGE_SIZE = 10
 const PAGE_COUNT = TOTAL_ROWS / PAGE_SIZE
 
-export const PoolTransactionsV3: FC<PoolTransactionsV3Props> = ({ pool, poolId }) => {
-  const [type, setType] = useState<Parameters<typeof useTransactionsV3>['2']['type']>(TransactionType.Swap)
+export const Transactions: FC<TransactionsProps> = ({ pool, poolId }) => {
+  const [type, setType] = useState<Parameters<typeof useTransactions>['2']['type']>(TransactionType.Swap)
   const [pageIndex, setPageIndex] = useState<number>(0)
 
   const COLUMNS = useMemo(() => {
@@ -38,7 +38,7 @@ export const PoolTransactionsV3: FC<PoolTransactionsV3Props> = ({ pool, poolId }
     [pageIndex, type]
   )
 
-  const { data, isLoading } = useTransactionsV3(pool, poolId, opts)
+  const { data, isLoading } = useTransactions(pool, poolId, opts)
 
   const dataPaginated = useMemo(
     () => (data ? data.slice(pageIndex * PAGE_SIZE, (pageIndex + 1) * PAGE_SIZE) : []),
@@ -56,7 +56,7 @@ export const PoolTransactionsV3: FC<PoolTransactionsV3Props> = ({ pool, poolId }
 
   return (
     <div>
-      <RadioGroup value={type} onChange={setType} className="flex gap-1 mb-6 justify-end px-2">
+      <RadioGroup value={type} onChange={setType} className="flex justify-end gap-1 px-2 mb-6">
         <RadioGroup.Option value={TransactionType.Swap}>
           {({ checked }) => (
             <Toggle size="sm" pressed={checked}>
@@ -92,7 +92,6 @@ export const PoolTransactionsV3: FC<PoolTransactionsV3Props> = ({ pool, poolId }
         page={pageIndex}
         hasPrev={pageIndex > 0}
         onNext={() => setPageIndex(pageIndex + 1)}
-        onPage={() => {}}
         hasNext={pageIndex < PAGE_COUNT - 1 && !!data && data?.length > (pageIndex + 1) * PAGE_SIZE}
         onPrev={() => setPageIndex(pageIndex - 1)}
         pageSize={PAGE_SIZE}
