+++ apps/evm/ui/pool/PoolSection/V3/Tabs/Statistics/Transactions/index.ts
@@ -1,0 +1,1 @@
+export * from './Transactions'
